Little fix Added in TaskList functionality to move Task Items to another TaskList

+++ TaskListApp/Services/TaskListService/TaskListService.cs
@@ -2,6 +2,7 @@
 using TaskListApp.Commands.TaskListCommands;
 using TaskListApp.Database.DBConnector;
 using TaskListApp.Database.Models.TaskListModel;
+using TaskListApp.Database.Models.TaskModels;
 using TaskListApp.Queries.TaskListQueries;
 
 namespace TaskListApp.Services.TaskListService
@@ -9,12 +10,10 @@ namespace TaskListApp.Services.TaskListService
     public class TaskListService : ITaskListService
     {
         private readonly ApplicationDbContext _context;
-        private readonly AuthenticationService _authenticationService;
 
-        public TaskListService(ApplicationDbContext context, AuthenticationService authenticationService)
+        public TaskListService(ApplicationDbContext context)
         {
             _context = context;
-            _authenticationService = authenticationService;
         }
 
         public async Task<TaskList> CreateTaskListAsync(CreateTaskListCommand command)
@@ -34,11 +33,18 @@ namespace TaskListApp.Services.TaskListService
 
         public async Task<TaskList> DeleteTaskListAsync(DeleteTaskListCommand command)
         {
-            var taskList = await _context.TaskLists.FindAsync(command.Id);
+            var taskList = await _context.TaskLists
+                .Include(list => list.Tasks)
+                .FirstOrDefaultAsync(list => list.Id == command.Id);
 
             if (taskList == null)
             {
-                throw new Exception("Список задач не найден");
+                throw new Exception("Task list not found");
+            }
+
+            if (taskList.Tasks.Any())
+            {
+                throw new Exception("Cannot delete a non-empty task list. Please use the method to delete a non-empty list and specify where tasks will be moved.");
             }
 
             _context.TaskLists.Remove(taskList);
@@ -55,7 +61,7 @@ namespace TaskListApp.Services.TaskListService
 
             if (taskList == null)
             {
-                throw new Exception("Список задач не найден"); 
+                throw new Exception("Task list not found");
             }
 
             return taskList;
@@ -67,7 +73,7 @@ namespace TaskListApp.Services.TaskListService
 
             if (taskList == null)
             {
-                throw new Exception("Список задач не найден");
+                throw new Exception("Task list not found");
             }
 
             taskList.Title = command.Title;
@@ -76,6 +82,46 @@ namespace TaskListApp.Services.TaskListService
             await _context.SaveChangesAsync();
 
             return taskList;
+        }
+
+        public async Task<TaskList> MoveTasksToAnotherList(MoveTasksToAnotherListCommand command)
+        {
+            var sourceList = await _context.TaskLists
+                .Include(list => list.Tasks)
+                .FirstOrDefaultAsync(list => list.Id == command.SourceListId);
+
+            if (sourceList == null)
+            {
+                throw new Exception("Source list not found");
+            }
+
+            if (!sourceList.Tasks.Any())
+            {
+                throw new Exception("Source list is empty.");
+            }
+
+            if (sourceList.Tasks.Any())
+            {
+                var targetList = await _context.TaskLists
+                    .Include(list => list.Tasks)
+                    .FirstOrDefaultAsync(list => list.Id == command.TargetListId);
+
+                if (targetList == null)
+                {
+                    throw new Exception("Target list not found");
+                }
+
+                foreach (var task in sourceList.Tasks)
+                {
+                    task.TaskListId = command.TargetListId;
+                }
+
+                await _context.SaveChangesAsync();
+
+                return targetList;
+            }
+
+            return null;
         }
     }
 }
